requirements.txt
kivy
openai
python-dotenv
SpeechRecognition
plyer
pyaudio
requests
beautifulsoup4
pypdf
python-docx